pyproject.toml
[tool.poetry]
name = "faust-download-server"
version = "0.1.0"
description = ""
authors = ["Thorsten Vitt <thorsten.vitt@uni-wuerzburg.de>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.78.0"
aiofiles = "^0.8.0"
uvicorn = {version = "^0.17.6", extras = ["standard"]}
zipstream-ng = "^1.3.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

